refactor: 抽离 WebSocketLifecycle（E2 Step 1a）
